Issue #64
Move date handling methods to their own class.
