+++ date_manager.py
@@ -1,0 +1,70 @@
+import pandas as pd
+from pandas.tseries.offsets import CustomBusinessDay
+from pandas.tseries.holiday import USFederalHolidayCalendar
+
+class DateManager:
+    def __init__(self):
+        cal = USFederalHolidayCalendar()
+        holidays = cal.holidays()
+        self.cbd = CustomBusinessDay(holidays=holidays)
+
+    def past_business_day(self, reference_date, business_days_past):
+        """
+        Calculates the business date a specified number of business days in the past,
+        skipping US federal holidays.
+
+        Parameters
+        ----------
+        reference_date : pd.Timestamp
+            Reference date to calculate days from.
+
+        business_days_past : int
+            The number of business days in the past.
+
+        Returns
+        -------
+        pd.Timestamp
+            The calculated past business date.
+        """
+        return reference_date - (self.cbd * business_days_past)
+
+    def future_business_day(self, reference_date, business_days_ahead):
+        """
+        Calculates the business date a specified number of business days in the future,
+        skipping US federal holidays.
+
+        Parameters
+        ----------
+        reference_date : pd.Timestamp
+            Reference date to calculate days from.
+
+        business_days_ahead : int
+            The number of business days in the future.
+
+        Returns
+        -------
+        pd.Timestamp
+            The calculated future business date.
+        """
+        return reference_date + (self.cbd * business_days_ahead)
+
+    def create_business_day_range(self, reference_date, num_days):
+        """
+        Creates a range of business days starting from a given date,
+        skipping US federal holidays.
+
+        Parameters
+        ----------
+        start_date : pd.Timestamp
+            The starting date
+
+        num_days : int
+            The number of business days to generate.
+
+        Returns
+        -------
+        pd.DatetimeIndex
+            A DatetimeIndex containing the range of business days.
+        """
+        return pd.date_range(start=reference_date, periods=num_days, freq=self.cbd)
+
